+++ R/proxy.R
@@ -22,13 +22,43 @@
 #' @param diag if \code{TRUE}, only compute diagonal elements of the
 #'   similarity/distance matrix; useful when comparing corresponding rows or
 #'   columns of `x` and `y`.
-#' @param use_nan if \code{TRUE}, return `NaN` if the standard deviation of a
+#' @param use_nan if `TRUE`, return `NaN` if the standard deviation of a
 #'   vector is zero when `method` is "correlation"; if all the values are zero
-#'   in a vector when `method` is "cosine", "kullback", "jeffreys" or "chisquared".
-#'   Note that use of `NaN` makes the similarity/distance matrix denser and
-#'   therefore larger.
+#'   in a vector when `method` is "cosine", "chisquared", "kullback", "jeffreys"
+#'   or "jensen". Note that use of `NaN` makes the similarity/distance matrix
+#'   denser and therefore larger in RAM. If `FALSE`, return zero in same use
+#'   situations as above. If `NULL`, will also return zero but also generate
+#'   a warning (default).
 #' @param digits determines rounding of small values towards zero. Use primarily
 #'   to correct rounding errors in C++. See \link{zapsmall}.
+#' @details
+#' Available methods for similarity:
+#' \itemize{
+#'   \item `cosine`: cosine similarity
+#'   \item `correlation`: Pearson's correlation
+#'   \item `jaccard`: Jaccard coefficient
+#'   \item `ejaccard`: the real value version of `jaccard`
+#'   \item `dice`: Dice coefficient
+#'   \item `edice`: the real value version of `dice`
+#'   \item `hamann`: Hamann similarity
+#'   \item `faith`: Faith similarity
+#'   \item `simple matching`: the percentage of common elements
+#' }
+#' Available methods for distance:
+#' \itemize{
+#'   \item `euclidean`: Euclidean distance
+#'   \item `chisquared`: chi-squared distance
+#'   \item `kullback`: Kullback–Leibler divergence
+#'   \item `jeffreys`: Jeffreys divergence
+#'   \item `jensen`: Jensen–Shannon divergence
+#'   \item `manhattan`: Manhattan distance
+#'   \item `maximum`: the largest difference between values
+#'   \item `canberra`: Canberra distance
+#'   \item `minkowski`: Minkowski distance
+#'   \item `hamming`: Hamming distance
+#' }
+#' See the vignette for how the similarity and distance are computed:
+#' `vignette("measures", package = "proxyC")`
 #' @import methods Matrix
 #' @importFrom RcppParallel RcppParallelLibs
 #' @seealso zapsmall
@@ -38,9 +68,9 @@
 #' simil(mt, method = "cosine")[1:5, 1:5]
 simil <- function(x, y = NULL, margin = 1,
                   method = c("cosine", "correlation", "jaccard", "ejaccard", "fjaccard",
-                             "dice", "edice", "hamann", "simple matching", "faith"),
+                             "dice", "edice", "hamann", "faith", "simple matching"),
                   min_simil = NULL, rank = NULL, drop0 = FALSE, diag = FALSE,
-                  use_nan = FALSE, digits = 14) {
+                  use_nan = NULL, digits = 14) {
 
     method[method == "hamman"] <- "hamann" # for transition
     method <- match.arg(method)
@@ -51,15 +81,16 @@ simil <- function(x, y = NULL, margin = 1,
 
 #' @rdname simil
 #' @param smooth adds a  fixed value to all the cells to avoid division by zero.
-#'   Only used when `method` is "chisquared", "kullback" or "jeffreys".
+#'   Only used when `method` is "chisquared", "kullback", "jeffreys" or "jensen".
+
 #' @export
 #' @examples
 #' mt <- Matrix::rsparsematrix(100, 100, 0.01)
 #' dist(mt, method = "euclidean")[1:5, 1:5]
 dist <- function(x, y = NULL, margin = 1,
-                 method = c("euclidean", "chisquared", "kullback", "jeffreys",
+                 method = c("euclidean", "chisquared", "kullback", "jeffreys", "jensen",
                             "manhattan", "maximum", "canberra", "minkowski", "hamming"),
-                 p = 2, smooth = 0, drop0 = FALSE, diag = FALSE, use_nan = FALSE, digits = 14) {
+                 p = 2, smooth = 0, drop0 = FALSE, diag = FALSE, use_nan = NULL, digits = 14) {
 
     method <- match.arg(method)
     proxy(x, y, margin, method, p = p, smooth = smooth, drop0 = drop0,
@@ -71,21 +102,22 @@ dist <- function(x, y = NULL, margin = 1,
 proxy <- function(x, y = NULL, margin = 1,
                   method = c("cosine", "correlation", "jaccard", "ejaccard", "fjaccard",
                              "dice", "edice", "hamann", "simple matching", "faith",
-                             "euclidean", "chisquared", "kullback", "jeffreys",
+                             "euclidean", "chisquared", "kullback", "jeffreys", "jensen",
                              "manhattan", "maximum", "canberra", "minkowski", "hamming"),
                   p = 2, smooth = 0, min_proxy = NULL, rank = NULL, drop0 = FALSE,
-                  diag = FALSE, use_nan = FALSE, digits = 14) {
+                  diag = FALSE, use_nan = NULL, digits = 14) {
 
     method[method == "hamman"] <- "hamann" # for transition
     method <- match.arg(method)
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
-
+    #x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(as(x, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
     symm <- is.null(y)
 
     if (is.null(y)) {
         y <- x
     } else {
-        y <- as(as(y, "CsparseMatrix"), "dgCMatrix")
+        #y <- as(as(y, "CsparseMatrix"), "dgCMatrix")
+        y <- as(as(as(y, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
     }
     if (!margin %in% c(1, 2))
         stop("Matrgin must be 1 (row) or 2 (column)")
@@ -104,14 +136,21 @@ proxy <- function(x, y = NULL, margin = 1,
         rank <- ncol(x)
     if (rank < 1)
         stop("rank must be great than or equal to 1")
-    if (!use_nan) {
-        if (method == "correlation") {
-            if (any(colSds(x) == 0) || any(colSds(y) == 0))
-                warning("x or y has vectors with zero standard deviation; consider setting use_nan = TRUE", call. = FALSE)
-        } else if (method %in% c("cosine", "kullback", "chisquared", "jeffreys")) {
-            if (any(colZeros(x) == nrow(x)) || any(colZeros(y) == nrow(y)))
-                warning("x or y has vectors with all zero; consider setting use_nan = TRUE", call. = FALSE)
+    if (is.null(use_nan)) {
+        if (method == "correlation" && (any(colSds(x) == 0) || any(colSds(y) == 0))) {
+            warning(paste0(
+                "x or y has vectors with zero standard deviation; ",
+                "consider setting use_nan = TRUE to set these values to NaN ",
+                "or use_nan = FALSE to suppress this warning"), call. = FALSE)
+        } else if (
+            method %in% c("cosine", "kullback", "chisquared", "jeffreys", "jensen") &&
+            (any(colZeros(x) == nrow(x)) || any(colZeros(y) == nrow(y)))) {
+            warning(paste0(
+                "x or y has vectors with all zero; ",
+                "consider setting use_nan = TRUE to set these values to NaN ",
+                "or use_nan = FALSE to suppress this warning"), call. = FALSE)
         }
+        use_nan <- FALSE
     }
     boolean <- FALSE
     weight <- 1
@@ -137,8 +176,8 @@ proxy <- function(x, y = NULL, margin = 1,
         weight <- p
     }
     if (boolean) {
-        x <- as(as(x, "lgCMatrix"), "dgCMatrix")
-        y <- as(as(y, "lgCMatrix"), "dgCMatrix")
+        x <- as(as(x, "lMatrix"), "dMatrix")
+        y <- as(as(y, "lMatrix"), "dMatrix")
     }
     if (method %in% c("cosine", "correlation", "euclidean") && !diag) {
         result <- cpp_linear(
@@ -159,7 +198,7 @@ proxy <- function(x, y = NULL, margin = 1,
                                      "hamann", "simple matching", "faith",
                                      "euclidean", "chisquared", "kullback", "manhattan",
                                      "maximum", "canberra", "minkowski", "hamming",
-                                     "jeffreys", "fjaccard")),
+                                     "jeffreys", "jensen", "fjaccard")),
             rank = rank,
             limit = min_proxy,
             weight = weight,
@@ -189,7 +228,7 @@ proxy <- function(x, y = NULL, margin = 1,
 #' apply(mt, 2, sd) # the same
 #' @export
 colSds <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- cpp_sd(x)
     names(result) <- colnames(x)
     return(result)
@@ -198,7 +237,7 @@ colSds <- function(x) {
 #' @rdname colSds
 #' @export
 rowSds <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- cpp_sd(t(x))
     names(result) <- rownames(x)
     return(result)
@@ -214,7 +253,7 @@ rowSds <- function(x) {
 #' apply(mt, 2, function(x) sum(x == 0)) # the same
 #' @export
 colZeros <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- nrow(x) - cpp_nz(x)
     names(result) <- colnames(x)
     return(result)
@@ -223,7 +262,7 @@ colZeros <- function(x) {
 #' @rdname colZeros
 #' @export
 rowZeros <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- ncol(x) - cpp_nz(t(x))
     names(result) <- rownames(x)
     return(result)